% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhist.R
\name{fhist}
\alias{fhist}
\title{Plot empirical distribution of a variable (histogram without binning)}
\usage{
fhist(x, freq = TRUE, col = "dodgerblue", lwd = 9, value.labels = TRUE, ...)
}
\arguments{
\item{x}{A numeric vector of values to plot frequencies for.}

\item{col}{Color for the frequency bars. Default is "dodgerblue".}

\item{lwd}{Line width for the frequency bars. Default is 9.}

\item{value.labels}{Logical. If TRUE, displays frequencies on top of each line. Default is TRUE}

\item{...}{Pass on any other argument that's accepted by \code{plot()}.}
}
\value{
Invisibly returns a data frame with values and their frequencies.
}
\description{
Creates a frequency plot showing the frequency of every observed value,
displaying the full range from minimum to maximum value.
}
\details{
This function creates a frequency plot where each observed value is shown
with its frequency. Unlike a standard histogram, there is no binning, unlike
a barplot, non-observed values of the variable are shown with 0 frequency 
instead of skipped.
}
\examples{
# Simple example
x <- c(1, 1, 2, 2, 2, 5, 5)
fhist(x)

# Pass on some common \code{plot()} arguments
fhist(x, col = "steelblue", xlab = "Value", ylab = "Frequency",ylim=c(0,7))

}
