% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2.R
\name{list2}
\alias{list2}
\title{Improves list() by allowing not naming (some) objects.}
\usage{
list2(...)
}
\arguments{
\item{...}{Objects to include in the list. Can be named or unnamed.}
}
\value{
A list where unnamed objects are automatically named with their existing name.
}
\description{
Improves list() by allowing not naming (some) objects.
}
\details{
This function is based on a solution on Stack Overflow:
\url{https://stackoverflow.com/questions/16951080/can-lists-be-created-that-name-themselves-based-on-input-object-names}

If objects are passed with explicit names (e.g., \code{list2(a = x, b = y)}),
those names are used. If objects are passed without names, the function
attempts to use the variable names of the objects themselves.
}
\examples{
x <- 1:5
y <- letters[1:3]
z <- matrix(1:4, nrow = 2)

# Create named list from objects
my_list <- list2(x, y, z)
names(my_list)  # "x" "y" "z"

# Works with explicit names too
my_list2 <- list2(a = x, b = y)
names(my_list2)  # "a" "b"

}
