% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.col.R
\name{message.col}
\alias{message.col}
\title{Colored Message Function}
\usage{
message.col(
  ...,
  col = "cyan",
  font = 1,
  domain = NULL,
  appendLF = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{...}{One or more R objects that can be coerced to character and pasted together.}

\item{col}{Character string specifying the color. Can be any color name (e.g., "cyan", 
"red", "blue", "dodgerblue", etc.). Default is "cyan". For ANSI terminal colors, 
standard colors (black, red, green, yellow, blue, magenta, cyan, white) will be used 
directly; other colors will default to "cyan".}

\item{font}{Integer. 1 for plain text (default), 2 for bold text.}

\item{domain}{See \code{\link[base]{gettext}}. If \code{NA}, messages will not be translated.}

\item{appendLF}{Logical. Should a newline be appended to the message?}

\item{quiet}{Logical. If TRUE, the function returns invisibly without printing.}
}
\description{
Prints a colored message with optional bold formatting.
}
\details{
This function prints colored messages to the console. If ANSI color codes are supported
by the terminal, the message will be colored. Otherwise, it will be printed as plain text.
}
\examples{
message.col("This is a plain cyan message", col = "cyan", font = 1)
message.col("This is a bold cyan message", col = "cyan", font = 2)
message.col("This is a bold red message", col = "red", font = 2)

}
