% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm2.R
\name{print.lm2}
\alias{print.lm2}
\title{Print method for lm2 objects}
\usage{
\method{print}{lm2}(x, notes, ...)
}
\arguments{
\item{x}{An object of class \code{lm2}}

\item{notes}{\code{deprecated}. Ignored. Use \code{\link{lm2_notes}} after printing.}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the original object
}
\description{
Print method for lm2 objects
}
