% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_simplified_table.R
\name{print.simplified_table}
\alias{print.simplified_table}
\title{Print method for simplified table output}
\usage{
\method{print}{simplified_table}(x, ...)
}
\arguments{
\item{x}{An object of class \code{simplified_table}}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for simplified table output
}
