% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.test2.R
\name{print.t.test2.enhanced}
\alias{print.t.test2.enhanced}
\title{Print method for enhanced t-test}
\usage{
\method{print}{t.test2.enhanced}(x, ...)
}
\arguments{
\item{x}{An object of class \code{t.test2.enhanced}}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for enhanced t-test
}
