% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table2.R
\name{print.table2}
\alias{print.table2}
\title{Print method for table2 output with centered column variable name}
\usage{
\method{print}{table2}(x, ...)
}
\arguments{
\item{x}{An object of class \code{table2}}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for table2 output with centered column variable name
}
