% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify.default}
\alias{simplify.default}
\title{Default method for simplify}
\usage{
\method{simplify}{default}(object, digits = 3, ...)
}
\arguments{
\item{object}{The object to simplify}

\item{digits}{Number of decimal places}

\item{...}{Additional arguments}
}
\description{
Default method for simplify
}
