% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_stimulus_cache.R
\name{clear_stimulus_cache}
\alias{clear_stimulus_cache}
\title{Clear cache used by stimulus.plot() to reuse resampling results from identical previous calls.}
\usage{
clear_stimulus_cache()
}
\description{
When `stimulus.plot()` with `plot.type = 'effects'` is run, resamples are saved in
package state. If an identical call is run again (same data, dv, condition, number of
simulations, etc.), stored results are loaded instead of re-calculated. Force
recalculation by clearing the cache with this function.
}
