% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_t.test2.R
\name{print.t.test2}
\alias{print.t.test2}
\title{Print method for t.test2 output}
\usage{
\method{print}{t.test2}(x, quiet = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{t.test2}}

\item{quiet}{Logical. If TRUE, suppresses missing value messages. Default is FALSE.}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for t.test2 output
}
